feat: NW Digital Twin 빌딩·장비실 이미지 텍스처 및 UI 개선
[장비실]
- Canvas resize 버그 수정: room→building 복귀 시 visibility:hidden 방식으로 Canvas 항상 마운트 유지
- 장비 이미지 텍스처화: server-blade.jpg 적용 (Textured Box 방식)
- 배경: scene.background 콜백 로드 방식 (alpha:false 유지, Bloom 충돌 해소)
- 모듈 레벨 프리로드 + 캐시로 배경 로딩 지연 개선
- 바닥 MeshReflectorMaterial 복구
- RACK 엣지 강화 (어두운 fill + lineWidth 1.4), RACK# 상단 라벨 제거

[빌딩]
- 배경 server-room-bg4.jpg, 층 텍스처 server-room-bg6.jpg 적용
- 바닥 면적 확장 (BW+30 × BD+30), 그리드 라인 제거, MeshReflectorMaterial 강화
- 빌딩 깊이(BD) 3.0→6.0, 폭(BW) 7.0→9.5, 층 높이(FH) 0.35→0.75
- 층 fill 투명도 제거 → 솔리드 텍스처
- 앞면 유리창 엣지 4칸 추가 (BW*0.10 균일 크기)
- 활성 층(8F·9F)만 라벨 표시, 폰트 15px, 피뢰침 높이 축소
- 엣지 색상 밝게(#7de8ff), lineWidth 1.2, 조명 전반 상향

[이미지 추가]
- public/images/: server-room-bg{1~6}, server-blade, server-blade-light, server-blade-steel, building-texture

+++ nttpoc_backend/src/main/java/com/mobigen/aiop/nttpoc/module/menu/dto/MenuResponse.java
@@ -3,8 +3,8 @@ package com.mobigen.aiop.nttpoc.module.menu.dto;
 import java.util.List;
 
 public record MenuResponse(
-        int menuId,
-        Integer parentId,
+        long menuId,
+        Long parentId,
         String menuName,
         String menuPath,
         String icon,
